Add Cache Hit Output
Cache hit was being used as stated in this pr
https://github.com/actions/setup-java/pull/275.
However, it was not included in the outputs of the action. It did not
cause any mayor issues, but if you were to lint the workflows, then an
output not defined error would pop up.
This Commit just adds the output to the list. Nothing more