Update README.md
A confusion introduced by running `java HelloWorldApp.java`
I believe the file is not an artifact of `setup-java`, thus user should provide it. But there is no tips in the guide that user should prepare it in advance. Directly copy the code snippets will introduce a straightforward error, which should not happen in a guide.